requirements.txt
pyyaml>=6.0
requests>=2.25
openai>=1.0